panderm_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.16
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
